as2-app-service/requirements.txt
Flask
waitress
requests
configparser
schedule
psycopg2
Flask-SQLAlchemy
Flask-Migrate
redis